+++ ui/src/Components/Search.js
@@ -1,6 +1,8 @@
 import React, { useState } from "react";
 
 export function SearchInput({ onSearch, placeholder }) {
+    console.log("onSearch is a function:", typeof onSearch === "function");
+
     const [keyword, setKeyword] = useState("");
 
     const handleChange = (event) => {
